Removed `@covers` annotations.
The scope that is covered by test methods is likely to change in the future,
meaning that the `@covers` annotation has to be changed too every time.
Coverage analysis does not require this annotation for coverage reports,
and will work just as well without it. Currently, there is no reason to limit
any particular test method's coverage, which is what the purpose of the
`@covers` annotation is; even more there is no reason to limit all methods'
coverage. Many developers apply an "all or nothing" rule when they decide
whether to place `@covers` annotations on particular test methods, and this
makes sense.

Therefore, this annotation has no real benefit here, while likely adding
more work to be done in the future.